[Refactoring ⚙️] delete Post hook제작 및 연결
